renamed game of life files
