+++ examples/game_of_life/life.c
@@ -1,230 +1,269 @@
 /*
-  life.c - written by Steve Furber - Aug 2013
+  life.c - written by Steve Furber and Alan Stokes - April 2015
 
   This example program demonstrates the use of Tubogrid to get simple
-  per-core animation. It implements Conway's Life on an XMAX x YMAX
-  chip array grid with 4*XMAX x 4*YMAX cells.
-
-  The only tricky part of the program is setting up the Router tables
-  - everything thereafter is pretty simple!
-  Each core sends a packet to each of its 8 neighbours every time it
-  recomputes its state (whether or not the state changes).
-  The update time is skewed from core to core to reduce SDP packet
-  contention, so a generation "parity" bit is used to keep the overall
-  model synchronous.
-  Connections to on-chip neighbours are direct, but when a core is on
-  the edge of the 4x4 array on a chip it must communicate with neighbours
-  on one or more other chips. The core routes its packet to horizontal
-  and vertical neighbours as required, where the corresponding core
-  sets up the route to the relevant cores on that chip.
-  For example, a packet from core (3,3) on chip (0,0) needs to go to:
-  - cores (2,2), (3,2) and (2,3) on chip(0.0) [core(3,3) on chip(0,0)]
-  - cores (2,0) and (3,0) on chip(0,1)        [core(3,3) on chip(0,1)]
-  - cores (0,2) and (0,3) on chip(1,0)        [core(3,3) on chip(1,0)]
-  - core (0,0) on chip(1,1)                   [core(3,3) on chip(1,1)]
-  Here the core responsible for local routing is indicated in [..].
-  Note that the diagonal connection to chip(1,1) is implemented via
-  chip(1,0) - the packet is routed horizontally and then vertically as
-  not all diagonal paths are available, so none are used.
+  per-core animation. It implements Conway's Life .
 */
+*******/
 
-#include <spin1_api.h>
+//! imports
+#include "spin1_api.h"
+#include "common-typedefs.h"
+#include <data_specification.h>
+#include <simulation.h>
+#include <debug.h>
 
-#define XMAX 2							// chip array width in chips
-#define YMAX 2							// chip array height in chips
-#define TOROID FALSE						// chip array = toroid?
+/*! multicast routing keys to communicate with neighbours */
+uint my_key;
+/*! setter for if i should record data to sdram*/
+uint should_record_to_sdram;
 
-#define EAST 1							// inter-chip link definitions
-#define NORTH 4							// NB diagonal links not used!
-#define WEST 8
-#define SOUTH 32
+//! this cell's state
+uint alive;
+//! the number of alive neigubours needed to keep this cell alive.
+uint theshold;
+//! previous state
+uint last_alive = 2;
+//! live neighbour count [for each gen]
+uint count [2];
+//! generation (used for parity)
+uint gen;
+//! control vlaue, which says how many timer tics to run for before exiting
+static uint32_t simulation_ticks;
+//! the unqieu identifier of this model, so that it can tell if the data its
+//! reading is for itself.
+#define APPLICATION_MAGIC_NUMBER 0xABCE
 
-uint alive;							// this cell's state
-uint last_alive = 2;						// previous state
-uint count [2];							// live neighbour count [for each gen]
-uint gen;							// generation (used for parity)
-uint myKey;							// key for sending
+//! human readable definitions of each region in SDRAM
+typedef enum regions_e {
+    SYSTEM_REGION, TRANSMISSIONS, STATE_REGION, RECORDING_REGION,
+} regions_e;
 
+//! human readable definitions of each element in the transmission region
+typedef enum transmission_region_elements {
+    HAS_KEY, MY_KEY
+}transmission_region_elements;
 
-void tick_callback (uint ticks, uint dummy)
+//! human readable definitions of each element in the initial state region
+typedef enum initial_state_region_elements {
+    INITIAL_STATE, THESHOLD,
+}initial_state_region_elements;
+
+//! human readable definitions of each element in the recoridng region config
+typedef enum recording_region_config_elements {
+    SET_TO_RECORD
+}recording_region_config_elements;
+
+/****f* life.c/report_state
+*
+* SUMMARY
+*  This function is used to report current temp
+*
+* SYNOPSIS
+*  void report_state (uint ticks)
+*
+* SOURCE
+*/
+void report_state(uint ticks)
 {
-  if (ticks % 250 == (myKey & 255))				// Every 0.25 secs, time skewed for SDP
-  {
-    alive = (count[gen] | alive) == 3;				// Life automaton rule
+    // if the model is set to record, then locate the right address to
+    //write current temp into
+    if (should_record_to_sdram){
+         // Copy data into recording channel
+         address_t address = data_specification_get_data_address();
+         address_t my_recording_region_address = data_specification_get_region(
+              RECORDING_REGION, address);
+         address_t place_to_record =
+              my_recording_region_address + sizeof(uint) +
+              (sizeof(uint) * ticks);
+         memcpy(&place_to_record, gen+(alive<<1), 4);
+    }
+}
+
+/****f* life.c/receive_data
+*
+* SUMMARY
+*  This function is used as a callback for packet received events.
+* receives data from neighbours and updates the checklist
+*
+* SYNOPSIS
+*  void receive_data (uint key, uint payload)
+*
+* INPUTS
+*   uint key: packet routing key - provided by the RTS
+*   uint payload: packet payload - provided by the RTS
+*
+* SOURCE
+*/
+void receive_data(uint key, uint payload)
+{
+    sark.vcpu->user1++;
+    log_debug("the key ive recieved is %d\n", key);
+    // count[gen mod 2] += alive
+    count[data & 1] += data >> 1;
+}
+
+
+/****f* heat_demo.c/send_first_value
+*
+* SUMMARY
+*
+* SYNOPSIS
+*  void send_first_value (uint a, uint b)
+*
+* SOURCE
+*/
+void send_first_value(uint a, uint b)
+{
+    use(a);
+    use(b);
+    log_debug("sending out inital state\n");
+    /* send data to neighbours */
+    while(! spin1_send_mc_packet(my_key, gen+(alive<<1), WITH_PAYLOAD)){
+        spin1_delay_us(1);
+    }
+}
+
+
+/****f* life.c/update
+*
+* SUMMARY
+*
+* SYNOPSIS
+*  void update (uint ticks, uint b)
+*
+* SOURCE
+*/
+void update(uint ticks, uint b)
+{
+    use(b);
+    sark.vcpu->user0++;
+
+    log_debug("on tick %d", ticks);
+    // check that the run time hasnt already alapsed and thus needs to be killed
+    if (ticks == simulation_ticks){
+        log_info("Simulation complete.\n");
+        spin1_exit(0);
+        return;
+    }
+
+    alive = (count[gen] | alive) == theshold;				// Life automaton rule
     count[gen] = 0;						// clear count for next gen
     gen = !gen;							// onto next generation
 
-    spin1_send_mc_packet (myKey, gen+(alive<<1), WITH_PAYLOAD);	// send state to neighbours
+    spin1_send_mc_packet (my_key, gen+(alive<<1), WITH_PAYLOAD);	// send state to neighbours
 
-    if (alive != last_alive)
-      {
-	char *s = (alive) ? "white" : "black";			// Make a colour string
-	io_printf (IO_STD, "#%s;#fill;\n", s);        		// And print it
-	last_alive = alive;
-      }
-  }
+    if (alive != last_alive){
+        char *s = (alive) ? "white" : "black";			// Make a colour string
+        log_debug("#%s;#fill;\n", s);        		// And print it
+        last_alive = alive;
+    }
+
+    if(should_record_to_sdram){
+        report_state(ticks);
+    }
+}
+
+//! \brief this method is to catch strange behaviour
+//! \param[in] key: the key being recieved
+//! \param[in] uknown: second arg with no state. set to zero by deault
+void receive_data_void(uint key, uint unknown){
+    use(key);
+    use(unknown);
+    log_error("this should never ever be done\n");
+}
+
+static bool initialize(uint32_t *timer_period) {
+    log_info("Initialise: started\n");
+
+    // Get the address this core's DTCM data starts at from SRAM
+    address_t address = data_specification_get_data_address();
+
+    // Read the header
+    if (!data_specification_read_header(address)) {
+        log_error("failed to read the dataspec header");
+        return false;
+    }
+
+    // Get the timing details
+    address_t system_region = data_specification_get_region(
+        SYSTEM_REGION, address);
+    if (!simulation_read_timing_details(
+            system_region, APPLICATION_MAGIC_NUMBER, timer_period,
+            &simulation_ticks)) {
+        log_error("failed to read the system header");
+        return false;
+    }
+
+    // output message about length of time to run
+    log_debug("i plan to run for %d timer ticks", simulation_ticks);
+
+    // initlise transmission keys
+    address_t transmission_region_address =  data_specification_get_region(
+        TRANSMISSIONS, address);
+    if (transmission_region_address[HAS_KEY] == 1){
+        my_key = transmission_region_address[MY_KEY];
+        log_debug("my key is %d\n", my_key);
+    }
+    else {
+        log_error("this heat element cant effect anything, deduced as an error,"
+                  "please fix the application fabric and try again\n");
+        return false;
+    }
+
+    // read my initial state
+    address_t state_region_address =  data_specification_get_region(
+        STATE_REGION, address);
+    alive = state_region_address[INITIAL_STATE];
+    theshold = state_region_address[THESHOLD];
+    log_debug("my initial state is %d", alive);
+
+    // read if im recording
+    address_t my_recording_region_address = data_specification_get_region(
+        RECORDING_REGION, address);
+    should_record_to_sdram = my_recording_region_address[SET_TO_RECORD];
+    log_debug("if i should record is set to %d\n", should_record_to_sdram);
+    return true;
 }
 
 
-void packet_in (uint key, uint data)				// data = gen + 2*alive
+/****f* heat_demo.c/c_main
+*
+* SUMMARY
+*  This function is called at application start-up.
+*  It is used to register event callbacks and begin the simulation.
+*
+* SYNOPSIS
+*  int c_main()
+*
+* SOURCE
+*/
+void c_main()
 {
-  count [data & 1] += data >> 1;				// count[gen mod 2] += alive
-}
+    log_info("starting game of life cell\n");
+    // Load DTCM data
+    uint32_t timer_period;
 
+    // initialise the model
+    if (!initialize(&timer_period)){
+    	rt_error(RTE_API);
+    }
 
-uint toH []   = {WEST, EAST};					// edge directions
-uint toV []   = {SOUTH, NORTH};					// to & from, Horizontal & Vertical
-uint fromH [] = {EAST, WEST};
-uint fromV [] = {NORTH, SOUTH};
+    // set timer tick value to 1ms (in microseconds)
+    // slow down simulation to alow users to appreciate changes
+    log_debug("setting timer to execute every %d microseconds", timer_period);
+    spin1_set_timer_tick(timer_period);
 
-uint route = 0;							// route bit vector
-int x, y;							// core x, y coordinates
+    // register callbacks
+    spin1_callback_on(MCPL_PACKET_RECEIVED, receive_data, 0);
+    spin1_callback_on(MC_PACKET_RECEIVED, receive_data_void, 0);
+    spin1_callback_on(TIMER_TICK, update, 0);
 
+    // kick-start the update process
+    spin1_schedule_callback(send_first_value, 0, 0, 3);
 
-uint edge (uint z)						// check for core at edge of chip
-{
-  return (z == 0) || (z == 3);
-}
-
-
-void add_route_to_core (int i, int j)				// filters out off-chip neighbours
-{
-  if (i >- 1 && i < 4  && j >- 1 && j < 4			// check for end of row & col
-      && !(i == x && j == y))
-    route += 1<<(i+4*j+7);					// don't route to self
-}
-
-
-uint next_chip (uint chip, uint dir)				// find next chip, modulo array size
-{
-  uint cX = chip >> 8;
-  uint cY = chip & 255;
-
-  // do the modulo sums
-
-  if (dir & SOUTH)
-    cY = cY ? cY - 1 : YMAX - 1;
-  if (dir & NORTH)
-    cY = (cY + 1) % YMAX;
-  if (dir & WEST)
-    cX = cX ? cX - 1 : XMAX - 1;
-  if (dir & EAST)
-    cX = (cX + 1) % XMAX;
-
-  return (cX << 8) + cY;
-}
-
-
-void add_route_to_chip (uint chip, uint dir)			// dir specifies possible directions
-{
-  if (TOROID)
-    route += dir;
-  else
-  {
-    uint nc = next_chip (chip, dir);				// toroidal routing on...
-    uint cX  = chip >> 8;
-    uint cY  = chip & 255;			// ...non-toroidal system
-    uint ncX = nc >> 8;
-    uint ncY = nc & 255;			// uses default routing to get back
-
-    if (cX != ncX)
-      route += toH[ncX > cX];
-
-    if (cY != ncY)
-      route += toV[ncY > cY];
-  }
-}
-
-
-void set_mc_table_entry (uint chip, uint core)			// route from (chip, core)
-{
-  uint e = rtr_alloc (1);
-
-  if (e == 0)
-    rt_error (RTE_ABORT);
-
-  rtr_mc_set (e, (chip << 8) + core, 0xffffffff, route);
-  route = 0;							// clear route for next use
-}
-
-
-void set_up_routing_tables (uint chip, uint core)		// this is the tricky bit!
-{
-  x = (core - 1) & 3;						// core x, y coordinates
-  y = (core - 1) >> 2;
-
-// first set up routes for internally-generated packets
-
-  for (int i = x - 1; i < x + 2; i++)				// column left, same, right
-    for (int j = y - 1; j < y + 2; j++)
-      add_route_to_core (i, j);					// row below, same, above
-
-  if (edge (x))
-    add_route_to_chip (chip, toH[x & 1]);			// send off-chip where appropriate
-
-  if (edge (y))
-    add_route_to_chip (chip, toV[y & 1]);
-
-  set_mc_table_entry (chip, core);				// write local Router table entry
-
-// then handle incoming packets from other chips...
-// ...each core handles routes from corresponding cores on other chips
-
-  if (edge (y))							// top & bottom
-  {
-    for (int i = x - 1; i < x + 2; i++)				// route to opposite row
-      add_route_to_core (i, 3 - y);
-
-    if (edge (x))
-      add_route_to_chip (chip, toH[x & 1]);			// forward diagonal connection
-
-    set_mc_table_entry (next_chip (chip, fromV[y & 1]), core);
-  }
-
-  if (edge (x))							// left & right
-  {
-    for (int j = y - 1; j < y + 2; j++)				// route to opposite col
-      add_route_to_core (3 - x, j);
-
-    set_mc_table_entry (next_chip (chip, fromH[x & 1]), core);
-  }
-
-  if (edge (x) && edge (y))					// incoming diagonal connection
-  {
-    add_route_to_core (3 - x, 3 - y);				// receive from opposite corner
-    set_mc_table_entry (next_chip (chip, fromH[x & 1] + fromV[y & 1]), core);
-  }
-}
-
-
-void init_Life_state (uint chip, uint core)
-{
-  if (chip == 0 && (core == 5 || core == 6 || core == 7
-    || core == 11 || core == 14)) count[0] = 3;			// initial state
-}
-
-
-void c_main (void)
-{
-  uint chip = spin1_get_chip_id ();				// get chip ID
-  uint core = spin1_get_core_id ();				// ...& core ID
-  myKey = (chip << 8) + core;					// key for my output packets
-
-  if ((chip >> 8) < XMAX && (chip & 255) < YMAX)		// only start required cores
-  {
-    io_printf (IO_BUF, "Started core %d %d %d\n",
-	  	chip >> 8, chip & 255, core);			// signal core running
-
-    set_up_routing_tables (chip, core);				// configure routing
-    init_Life_state (chip, core);				// Initialise Life state
-
-    spin1_set_timer_tick (1000);				// 1ms timer tick
-    spin1_callback_on (TIMER_TICK, tick_callback, 1);		// timer callback
-    spin1_callback_on (MCPL_PACKET_RECEIVED, packet_in, -1);	// incoming packet callback
-
-    spin1_start (SYNC_NOWAIT);					// start event-driven operation
-
-    // print if event_exit used...
-    log_debug("Terminated core %d %d %d\n", chip >> 8, chip & 255, core);
-  }
+    // start execution
+    log_info("Starting\n");
+    simulation_run();
+    log_debug("ending game of life cell");
 }
